wip: move `checkPieceContentStatus` to server code
