+++ meteor/server/publications/pieceContentStatusUI/rundown/regenerateItems.ts
@@ -11,8 +11,8 @@ import { literal, protectString } from '../../../../lib/lib'
 import { CustomPublishCollection } from '../../../lib/customPublication'
 import { ContentCache } from './reactiveContentCache'
 import { checkPieceContentStatusAndDependencies, PieceDependencies, StudioMini } from '../common'
-import { PieceContentStatusPiece } from '../../../../lib/mediaObjects'
 import { wrapTranslatableMessageFromBlueprintsIfNotString } from '@sofie-automation/corelib/dist/TranslatableMessage'
+import { PieceContentStatusPiece } from '../checkPieceContentStatus'
 
 async function regenerateGenericPiece(
 	contentCache: ReadonlyDeep<ContentCache>,
